Rewrite the vagrant-illumos-gate wrappers.
Unify them into a common command, 'vig', whose supporting code lives in
vig-common.  Add better help/navigation for the commands.

Change the general mechanism to revolve around a single git repo that lives
on the host rather than also having a copy on the guest.  Save a copy of
vig/vig-common from the master branch into the guest, and provide a means
for updating that copy.

Dispose of the alternative recipes for OmniOS/SmartOS, since they're not
able to build illumos-gate.  OI hipster is the only illumos distribution
currently capable of this.  This can change later, but there is room in this
structure to allow that.

This change isn't quite complete; various bugfixes are needed.