Rework workspace & VM setup.
Always run workspace setup prior to executing a build; make the setup
routine a no-op if no changes are needed.  Store the checksum of the most
recently extracted closed binaries to enable this.  Perform the workspace
setup on the host rather than the guest to reduce time required.o

Separate VM setup from starting the VM, so the build step can just ensure
the VM has been started, regardless of whether it needs to be setup.