Revert shared-folder-only approach.
It's non-functional, so we'll have to stick to using a local clone inside
the VM.  Try to make this easier by automatically pulling updates into that
clone from the master and checking out branches as requested.  This will
make it harder to inspect build logs from the host etc, but it'll have to do.